Refactor: split healing and file-based healing
Logically the same, but the procedures are different.
If turns our we can always use file based diffing, it's easy
to remove the pipeline
