+++ jobs/healing/main.js
@@ -2,7 +2,8 @@ import { querySudo as query } from '@lblod/mu-auth-sudo';
 import { sparqlEscapeUri } from 'mu';
 import { appendTaskError, loadTask, updateTaskStatus } from '../../lib/task';
 import { parseResult, storeError } from '../../lib/utils';
-import { runHealingTask } from './pipeline-healing';
+import { runHealingTask as runHealingTaskOnFile } from './file-based-healing';
+import { runHealingTask } from './standard-healing';
 
 export async function executeHealingTask(service_config, service_export_config) {
 
@@ -19,7 +20,14 @@ export async function executeHealingTask(service_config, service_export_config) 
       const task = await loadTask(service_config, syncTaskUri || healingTaskUri);
       try {
         await updateTaskStatus(task, service_config.statusBusy);
-        await runHealingTask(service_config, service_export_config, task, syncTaskUri ? true : false, service_config.serveDeltaFiles && healingTaskUri);
+
+        if(service_config.useFileDiff) {
+          await runHealingTaskOnFile(service_config, service_export_config, task, syncTaskUri ? true : false, service_config.serveDeltaFiles && healingTaskUri);
+        }
+        else {
+          await runHealingTask(service_config, service_export_config, task, syncTaskUri ? true : false, service_config.serveDeltaFiles && healingTaskUri);
+        }
+
         await updateTaskStatus(task, service_config.statusSuccess);
       }
       catch(e) {
